requirements.txt
# Zone Crossing Minimization - Dependencies
# Install with: pip install -r requirements.txt

# ==== Core Dependencies ====
matplotlib>=3.5.0
numpy>=1.21.0
pyyaml>=6.0
rich>=13.0.0

# ==== PyTorch and Deep Learning ====
# CPU-only version (smaller, works everywhere):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
# GPU version (CUDA 11.8):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# GPU version (CUDA 12.1):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
torchvision>=0.15.0

# ==== Training Utilities ====
tensorboard>=2.12.0
tqdm>=4.65.0
